add supervisor and jade (provision)
Supervisor is for auto-reloading node when there's a change in .js files. Run with "supervisor app.js".

Jade is the HTML preprocessing template that Express uses.

3 node.js files added: app.js, routes/index.js and views/index.jade
where app.js is the main, routes/index.js is responsible for routing and views/index.jade will be used to generate the "index html" file.

Browser-sync will also be downloaded, but not implemented.